Add a calcWOperator() function to evaluate numbers
calcWOperator() utilizes handleNumberInput() along with operate() and
the four basic math operating functions to do the following: store
the numbers plus the operator and perform an operation.

The operation of this calculator does not follow operator precedence
like that in mathematics. It does not evaluate more than a single pair
of numbers at a time. For e.g., in the operation: 12 + 7 -, first 19
is calculated and then it is continued.